refactored ionq gates redundancies
