+++ qbraid/transpiler/conversions/qasm2/qasm2_to_ionq.py
@@ -15,13 +15,14 @@ Module containing functions to convert between OpenQASM 2 and IonQ JSON format.
 from __future__ import annotations
 
 import re
-from enum import Enum
 from typing import TYPE_CHECKING, Any, Union
 
 import openqasm3.ast
 
+from qbraid.passes.qasm.analyze import has_measurements
 from qbraid.passes.qasm.compat import convert_qasm_pi_to_decimal
 from qbraid.programs import load_program
+from qbraid.programs.gate_model.ionq import IonQProgram
 from qbraid.programs.gate_model.qasm2 import OpenQasm2Program
 from qbraid.programs.gate_model.qasm3 import OpenQasm3Program
 from qbraid.transpiler.annotations import weight
@@ -62,42 +63,6 @@ IONQ_THREE_QUBIT_GATE_MAP = {
     "toffoli": "cnot",
 }
 
-IONQ_NATIVE_BASE = {"gpi", "gpi2"}
-IONQ_NATIVE_GATES = IONQ_NATIVE_BASE | {"ms", "zz"}
-
-
-class GateSet(Enum):
-    """Enumeration of IonQ gate sets types, native and qis (abstract)."""
-
-    NATIVE = "native"
-    QIS = "qis"
-
-
-def determine_gateset(circuit: list[dict[str, Any]]) -> GateSet:
-    """Determines the gate set of an IonQ circuit gate list.
-
-    Args:
-        circuit (list[dict]): The IonQ circuit to analyze.
-
-    Returns:
-        GateSet: The gate set of the circuit.
-
-    Raises:
-        ValueError: If the circuit mixes native and abstract (qis) gates.
-    """
-    contains_gpi = any(instr["gate"] in IONQ_NATIVE_BASE for instr in circuit)
-
-    for instr in circuit:
-        gate = instr.get("gate")
-        if gate not in IONQ_NATIVE_GATES:
-            if contains_gpi:
-                raise ValueError(
-                    f"Invalid gate '{gate}' encountered with gpi/gpi2 present in the circuit."
-                )
-            return GateSet.QIS
-
-    return GateSet.NATIVE
-
 
 def qasm_to_ionq(qasm: QasmStringType) -> IonQDictType:
     """Returns an IonQ JSON format representation the input OpenQASM string.
@@ -108,6 +73,8 @@ def qasm_to_ionq(qasm: QasmStringType) -> IonQDictType:
     Returns:
         dict: IonQ JSON format equivalent to input OpenQASM string.
     """
+    if has_measurements(qasm):
+        raise ValueError("Circuits with measurements are not supported by the IonQDictType")
 
     qprogram: Union[OpenQasm2Program, OpenQasm3Program] = load_program(qasm)
     num_qubits = qprogram.num_qubits
@@ -176,7 +143,7 @@ def qasm_to_ionq(qasm: QasmStringType) -> IonQDictType:
             else:
                 raise NotImplementedError(f"'{name}' gate not yet supported")
 
-    gateset = determine_gateset(gates)
+    gateset = IonQProgram.determine_gateset(gates)
 
     return {"qubits": num_qubits, "circuit": gates, "gateset": gateset.value}
